// src/model/adress.model.ts
export class Adress {
  id: number;
  gender: string;
  adressName: string;
  lastName: string;
  firstName: string;
  phone: string;
  digicode: string;
  interphone: string;
  appartNumber: string;
  floor: number;
  building: string;
  number: number;
  street: string;
  postalCode: string;
  city: string;
  information: string;
  delivery: boolean;
  billing: boolean;
}